requirements.txt
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
Werkzeug==1.0.1
certifi==2020.11.8
chardet==3.0.4
idna==2.10
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
dnspython